p4i integration Integrate 101390 from 3 to main: Integrate 101389 from 3.2 to 3: Integrate 101388 from 3.1 to 3.2: Fix sign warning.
[git-p4: depot-paths = "//depot/qt/main/": change = 101391]

+++ src/kernel/qimage.cpp
@@ -5521,7 +5521,7 @@ static void write_xbm_image( QImageIO *iio )
 		*p++ = '\n';
 		*p++ = ' ';
 		*p   = '\0';
-		if ( qstrlen(buf) != d->writeBlock( buf, qstrlen(buf) ) ) {
+		if ( (int)qstrlen(buf) != d->writeBlock( buf, qstrlen(buf) ) ) {
 		    iio->setStatus( 1 );
 		    return;
 		}
@@ -5531,7 +5531,7 @@ static void write_xbm_image( QImageIO *iio )
 	}
     }
     strcpy( p, " };\n" );
-    if ( qstrlen(buf) != d->writeBlock( buf, qstrlen(buf) ) )
+    if ( (int)qstrlen(buf) != d->writeBlock( buf, qstrlen(buf) ) )
 	iio->setStatus( 1 );
 }
